Implemented delete authorization

+++ WS03/App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+
 
 class ListingController
 {
@@ -20,7 +22,7 @@ class ListingController
     {
 
 
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
 
         loadView(
@@ -81,7 +83,7 @@ class ListingController
 
         $newListingData = array_intersect_key($_POST, array_flip($allowfields));
 
-        $newListingData['user_id'] = 1; // hardcoded user id for now
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -150,6 +152,16 @@ class ListingController
             ErrorController::notFound('Listing Not Found');
             return;
         }
+
+        //authorization
+        if (Session::get('user')['id'] !== $listing->user_id) {
+            $_SESSION['error_message'] = 'You are not authorized to delete this listing';
+            return redirect('/listings');
+        }
+
+
+
+
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
